// src/app/api/model/tvEpisodeImages.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Image } from './image';


export interface TvEpisodeImages { 
    id?: number;
    stills?: Array<Image>;
}