// src/app/api/model/videoList.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Video } from './video';


export interface VideoList { 
    id?: number;
    results?: Array<Video>;
}